* improved data structure and io

+++ src/definitions.h
@@ -33,9 +33,9 @@ namespace ProMapAnalyzer {
     typedef float f32;
     typedef double f64;
 
-    typedef u32 vertex_t;
-    typedef s32 weight_t;
-    typedef u32 partition_t;
+    typedef u64 vertex_t;
+    typedef s64 weight_t;
+    typedef u64 partition_t;
 }
 
 #endif //PROCESSMAPPINGANALYZER_DEFINITIONS_H
